task-3/4 | Оптимизирована точка входа
